Switch cache count to long

+++ src/Headless.Caching.Redis/RedisCache.cs
@@ -633,7 +633,7 @@ public sealed class RedisCache(
         return await _database.KeyExistsAsync(_GetKey(key));
     }
 
-    public async ValueTask<int> GetCountAsync(string prefix = "", CancellationToken cancellationToken = default)
+    public async ValueTask<long> GetCountAsync(string prefix = "", CancellationToken cancellationToken = default)
     {
         cancellationToken.ThrowIfCancellationRequested();
